% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compmDistMat.R
\name{compmDistMat}
\alias{compmDistMat}
\title{compmDistMat}
\usage{
compmDistMat(
  x,
  y = NULL,
  method = "Euclidean",
  parallel = FALSE,
  cl = NULL,
  diag = TRUE,
  upper = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{list} of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Values of functions should be of the same time points.}

\item{y}{\code{list} of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{k}) and rows are discrete-time points (\eqn{t}). Values of functions should be of the same time points.}

\item{method}{\code{string} to specify the distance measure to compute. Defaults to "Euclidean". The complete list of available options is returned by \code{mdistMeasures()}.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{diag}{is a logical value indicating whether the diagonal of the distance matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, diagonal values are NA.}

\item{upper}{is a logical value indicating whether the upper triangle of the matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, upper triangle values are NA.}

\item{...}{other arguments.}
}
\value{
The present function returns a rectangular matrix of size \eqn{k*n} containing the distances of
all \eqn{k} functions in \code{y} with all n functions/trajectories in \code{x}. If \code{y} is left unspecified, the function returns the distances of the functions in \code{x} with itself (a symmetric \eqn{n x n} matrix).
}
\description{
Distance matrix among \eqn{m}-dimensional functions / trajectories
}
\details{
The present function computes the distance matrix of one set of \eqn{n} different \eqn{m}-dimensional functions to another set of \eqn{k} different \eqn{m}-dimensional functions. The distances for all combinations of functions in the first set with functions of the second set are returned.
Different distance measures are available and can be set by \code{method}.
}
\seealso{
See \code{\link[parallel:makeCluster]} \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}.
}
