% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiNestedCV.R
\name{innerResultsExtract}
\alias{innerResultsExtract}
\title{innerResultsExtract}
\usage{
innerResultsExtract(x, h)
}
\arguments{
\item{x}{\code{vector}}

\item{h}{\code{vector}}
}
\description{
innerResultsExtract
}
