% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelFunctions.R
\name{Ker.norm}
\alias{Ker.norm}
\title{Ker.norm
computes the values of the normal kernel. Caps values from below at 1e-40, so that no divisions by 0 occur}
\usage{
Ker.norm(x)
}
\arguments{
\item{x}{is the entry for Ker.norm function.}
}
\description{
Ker.norm
computes the values of the normal kernel. Caps values from below at 1e-40, so that no divisions by 0 occur
}
