% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mAitchison.R
\name{mAitchison}
\alias{mAitchison}
\title{Aitchison distance from package robCompositions}
\usage{
mAitchison(data)
}
\arguments{
\item{data}{is the object with the information related to the trajectories, i.e., number of times the mouse was in each of the areas of interest.}
}
\description{
Aitchison distance from package robCompositions
}
