% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiEnsemble.R
\name{mclassiEnsembleTest}
\alias{mclassiEnsembleTest}
\title{mclassiEnsembleTest}
\usage{
mclassiEnsembleTest(learner, model, task)
}
\arguments{
\item{learner:}{learner of class EnsembleLearner}

\item{model:}{model of class EnsembleTrain}

\item{task:}{define a task with makeClassifTask}
}
\description{
mclassiEnsembleTest
}
