% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiEnsemble.R
\name{mclassiEnsembleTrain}
\alias{mclassiEnsembleTrain}
\title{mclassiEnsembleTrain}
\usage{
mclassiEnsembleTrain(learner, task, subset, cv = 5, set.seed = NULL)
}
\arguments{
\item{learner:}{learner of class EnsembleLearner}

\item{task:}{define a task with makeClassifTask}

\item{subset:}{ith inner sample}

\item{cv:}{splits in the k-fold cross-validation in the inner loop. By default, 5-fold cross-validation is considered}

\item{set.seed:}{seed for the splits in the inner-loop}
}
\description{
mclassiEnsembleTrain
}
