% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiEnsemble.R
\name{mclassiGenStack}
\alias{mclassiGenStack}
\title{mclassiGenStack}
\usage{
mclassiGenStack(model, prediction, par.vals.ensemble, xreg = NULL)
}
\arguments{
\item{model:}{model of class mclassiEnsembleTest}

\item{pred:}{predictions of an object of class mclassiEnsembleTest}

\item{par.vals.ensemble:}{list including super.learner (randomForest or boosting), iters for the k-fold cross-validation
related to the super.learner, and par.vals.ensemble setting of parameter candidates to be tuned using k-fold cross-validation,
and set.seed for the splits of the k-fold cross-valdation to tune the super.learner}

\item{xreg:}{if additional covariates are considered. Default is NULL.}
}
\description{
mclassiGenStack
}
