% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiKernel.R
\name{mclassiKernel}
\alias{mclassiKernel}
\title{mclassiKernel}
\usage{
mclassiKernel(
  classes,
  fdata,
  mdist = NULL,
  metric = "Euclidean",
  kernel = "Ker.norm",
  h = 1,
  nderiv = 0L,
  parallel = FALSE,
  cl = NULL,
  diag = FALSE,
  upper = FALSE,
  ...
)
}
\arguments{
\item{classes}{\code{factor} or \code{numeric}. A vector containing the true classes of the training data.}

\item{fdata}{the training covariates as a \code{list} of \eqn{m} objects in \code{matrix} form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Values of functions should be of the same time points.}

\item{mdist}{\code{matrix}. A customized matrix of distances. Default is NULL.}

\item{metric}{\code{string} specifying the metric for knn. The complete list of available metrics is returned by
\code{mdistMeasures()}.}

\item{kernel}{\code{string} indicates the type of kernel function to be used. Default is Ker.norm.}

\item{h}{\code{numeric} bandwith parameter. Default is 1.}

\item{nderiv}{\code{integer} number of derivatives of the trajectory.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{diag}{is a logical value indicating whether the diagonal of the distance matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, diagonal values are NA.}

\item{upper}{is a logical value indicating whether the upper triangle of the matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, upper triangle values are NA.}

\item{...}{other arguments.}
}
\value{
Returns a list containing the parameters and the training data for a mknn prediction. This list is used for prediction with the function \code{\link[classiMultiFuncTest:predict.mclassiKnn]{predict.mclassiKnn}}
}
\description{
Create mclassiKernel Object
}
\details{
The present function creates an object to specify the kernel-based nonparametric
curves discrimination for \eqn{n}-different observations of
\eqn{m}-dimensional functionnal data.
}
\seealso{
See \code{\link[parallel:makeCluster]} \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}
}
