% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiEnsemble.R
\name{mclassiLC}
\alias{mclassiLC}
\title{mclassiCL}
\usage{
mclassiLC(model, prediction, par.vals.ensemble)
}
\arguments{
\item{model:}{model of class mclassiEnsembleTest}

\item{pred:}{predictions of an object of class mclassiEnsembleTest}

\item{par.vals.ensemble:}{parameters needed for least squares minimization with function lsei from package limSolve}
}
\description{
mclassiCL
}
