% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiTies.R
\name{mclassiTies}
\alias{mclassiTies}
\title{mclassiTies}
\usage{
mclassiTies(x, knn, set.seed = NULL, ...)
}
\arguments{
\item{x}{vector of distances of a new trajectory to all training trajectories.}

\item{knn}{number of nearest neighbours.}

\item{...}{other arguments.}

\item{if}{a seed wants to be specified.}
}
\description{
Provides the positions of the indices of the k nearest trajectories.
Selects randomly among ties for the \eqn{k}th nearest trajectory.
}
