% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdcor.R
\name{mdcor}
\alias{mdcor}
\title{mdcor}
\usage{
mdcor(data, parallel = FALSE, cl = NULL, index = 1)
}
\arguments{
\item{data}{a list of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Values of functions should be of the same time points.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}
}
\description{
mdcor distance between multivariate trajectories
}
\details{
Computes the dcor distance for all pairs of \eqn{m}-dimensional trajectories.
}
