% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdistMeasures.R
\name{mdistMeasures}
\alias{mdistMeasures}
\title{mdistMeasures}
\usage{
mdistMeasures(data, measure)
}
\arguments{
\item{data}{\code{list} containing one vector (or two if \code{measure="flips"})
of measures computed by the function \code{\link[classiMultiFunc:mmeasures]{mmeasures}}.}

\item{measure}{\code{string} to specify the measure contained in \code{data}.
The complete list of available measures can be retrieved with the function \code{\link[compmDistMat:mmeasuresChoices]{mmeasuresChoices}}.
For a detailed explanation of the different measures see \code{\link[mousetrap:mt_measures]{mousetrap:mt_measures}}.}
}
\description{
Distance between measures of 2-Dimensional trajectories.
}
\details{
The mousetrap package enables the calculation of certain measures of mouse trajectories.
The present function enables the computation of "distances" between such measures.
}
