% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtw.R
\name{mdtw}
\alias{mdtw}
\title{mdtw}
\usage{
mdtw(data, parallel = FALSE, cl = NULL, ...)
}
\arguments{
\item{data}{a list of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Values of functions should be of the same time points.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{...}{additional arguments that can be passed to \code{\link[dtw:dtw]{dtw::dtw}}}
}
\value{
Returns a square and symmetric \eqn{n x n} matrix of \eqn{m}-dimensional Dynamic Time Warp distances.
}
\description{
Dynamic Time Warping distance between multivariate functions
}
\details{
Computes the Dynamic Time Warping distance for all pairs of \eqn{m}-dimensional functions.

Applies the \code{\link[dtw:dtw]{dtw::dtw}} function  to all pairs of \eqn{m}-dimensional functions.
By default, the Euclidean Distance between optimally aligned \eqn{m}-dimensional functions is computed.
Other options and parameters that can be passed to \code{\link[dtw:dtw]{dtw::dtw}} (such as the use of the Manhattan distance) can be defined in the additional parameters \code{...}
}
\references{
\enumerate{
\item Toni Giorgino. \emph{Computing and Visualizing Dynamic Time
Warping Alignments in R: The dtw Package.} Journal of Statistical Software,
31(7), 1-24.  \ifelse{text}{doi:10.18637/jss.v031.i07 <https://doi.org/10.18637/jss.v031.i07>}{\ifelse{latex}{\href{https://doi.org/10.18637/jss.v031.i07}{doi:10.18637\out{\\slash{}}jss.v031.i07}}{\href{https://doi.org/10.18637/jss.v031.i07}{doi:10.18637/jss.v031.i07}}}
\item Tormene, P.;
Giorgino, T.; Quaglini, S. & Stefanelli, M. \emph{Matching incomplete time
series with dynamic time warping: an algorithm and an application to
post-stroke rehabilitation.} Artif Intell Med, 2009, 45, 11-34.
\ifelse{text}{doi:10.1016/j.artmed.2008.11.007 <https://doi.org/10.1016/j.artmed.2008.11.007>}{\ifelse{latex}{\href{https://doi.org/10.1016/j.artmed.2008.11.007}{doi:10.1016\out{\\slash{}}j.artmed.2008.11.007}}{\href{https://doi.org/10.1016/j.artmed.2008.11.007}{doi:10.1016/j.artmed.2008.11.007}}}
\item Sakoe, H.;
Chiba, S., \emph{Dynamic programming algorithm optimization for spoken word
recognition,} Acoustics, Speech, and Signal Processing,
IEEE Transactions on , vol.26, no.1, pp. 43-49, Feb 1978.
\ifelse{text}{doi:10.1109/TASSP.1978.1163055 <https://doi.org/10.1109/TASSP.1978.1163055>}{\ifelse{latex}{\href{https://doi.org/10.1109/TASSP.1978.1163055}{doi:10.1109\out{\\slash{}}TASSP.1978.1163055}}{\href{https://doi.org/10.1109/TASSP.1978.1163055}{doi:10.1109/TASSP.1978.1163055}}}
\item Mori, A.; Uchida, S.; Kurazume, R.; Taniguchi, R.; Hasegawa, T. & Sakoe, H.
\emph{Early Recognition and Prediction of Gestures} Proc. 18th International
Conference on Pattern Recognition ICPR 2006, 2006, 3, 560-563 \ifelse{text}{doi:10.1109/ICPR.2006.467 <https://doi.org/10.1109/ICPR.2006.467>}{\ifelse{latex}{\href{https://doi.org/10.1109/ICPR.2006.467}{doi:10.1109\out{\\slash{}}ICPR.2006.467}}{\href{https://doi.org/10.1109/ICPR.2006.467}{doi:10.1109/ICPR.2006.467}}}
\item Sakoe,
H. \emph{Two-level DP-matching--A dynamic programming-based pattern matching
algorithm for connected word recognition} Acoustics, Speech, and Signal
Processing, IEEE Transactions on, 1979, 27, 588-595 \ifelse{text}{doi:10.1109/TASSP.1979.1163310 <https://doi.org/10.1109/TASSP.1979.1163310>}{\ifelse{latex}{\href{https://doi.org/10.1109/TASSP.1979.1163310}{doi:10.1109\out{\\slash{}}TASSP.1979.1163310}}{\href{https://doi.org/10.1109/TASSP.1979.1163310}{doi:10.1109/TASSP.1979.1163310}}}
\item Rabiner L, Rosenberg A, Levinson
S (1978). \emph{Considerations in dynamic time warping algorithms for
discrete word recognition.} IEEE Trans. Acoust., Speech, Signal Process.,
26(6), 575-582.  \ifelse{text}{doi:10.1109/TASSP.1978.1163164 <https://doi.org/10.1109/TASSP.1978.1163164>}{\ifelse{latex}{\href{https://doi.org/10.1109/TASSP.1978.1163164}{doi:10.1109\out{\\slash{}}TASSP.1978.1163164}}{\href{https://doi.org/10.1109/TASSP.1978.1163164}{doi:10.1109/TASSP.1978.1163164}}}
\item Muller M. \emph{Dynamic Time
Warping} in \emph{Information Retrieval for Music and Motion}. Springer
Berlin Heidelberg; 2007. p. 69-84. \ifelse{text}{doi:10.1007/978-3-540-74048-3_4 <https://doi.org/10.1007/978-3-540-74048-3_4>}{\ifelse{latex}{\href{https://doi.org/10.1007/978-3-540-74048-3_4}{doi:10.1007\out{\\slash{}}978\out{\\-}3\out{\\-}540\out{\\-}74048\out{\\-}3_4}}{\href{https://doi.org/10.1007/978-3-540-74048-3_4}{doi:10.1007/978-3-540-74048-3_4}}}
}
}
\seealso{
See \code{\link[parallel:makeCluster]} \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}, \code{\link[proxy:dist]{dist}} from \href{https://cran.r-project.org/web/packages/proxy/index.html}{proxy}
}
