% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrangeU.R
\name{mrangeU}
\alias{mrangeU}
\title{mrangeU}
\usage{
mrangeU(data, parallel = FALSE, cl = NULL)
}
\description{
Range distance between multivariate trajectories
}
\details{
Computes the Range distance for all pairs of \eqn{m}-dimensional trajectories For a single pair of functions,
the current R function returns the absolute difference between the ranges of \eqn{m}-dimensional trajectories.

x = replicate ( 4 , rnorm ( 100 , 0 , 3 ) )
mrangeU ( x , parallel = FALSE , cl = NULL )
}
