% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mclassiKernel.R
\name{predict.mclassiKernel}
\alias{predict.mclassiKernel}
\title{predict.mclassiKernel}
\usage{
\method{predict}{mclassiKernel}(object, newdata = NULL, Mdist = NULL, predict.type = "response")
}
\arguments{
\item{object}{mclassiKernel object defined with \code{\link[classiMultiFuncTest:mclassiKernel]{mclassiKernel}}}

\item{newdata}{the functional covariates for prediction as a \code{list} of \eqn{m} objects in \code{matrix} form.
Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time
points (\eqn{t}). Functions' values should be of the same time points as the training data.}

\item{predict.type}{\code{string} to specify type of prediction. Either \code{"response"} or \code{"prob"}}
}
\description{
Predictions with an mclassiKernel object
}
\details{
The present function yields predictions for new data with an mclassiKernel object .
}
\examples{

## 3-dimensional functions
#classes
classes = as.factor ( c ( "Cat" , "Dog" , "Dog" , "Cat" ) )

#training_data
x = replicate ( 4 , rnorm ( 100 , 0 , 3 ) )
y = replicate ( 4 , rnorm ( 100 , 3 , 1 ) )
z = replicate ( 4 , rpois ( 100 , 2 ) )
training_data = data.frame ( cbind ( id = 1:4 , dim = 3 , t ( x ), t ( y ) , t ( z ) ) )  

#Creating object for mclassiKernel prediction
object = mclassiKernel ( classes = classes , fdata = training_data , h = 30 , kernel = "Ker.norm" , nderiv = 0 , cl = NULL )

#test_data
x = replicate (2, rnorm ( 100 , 0 , 2.9 ) )
y = replicate (2, rnorm ( 100 , 3 , 1.5 ) )
z = replicate (2, rpois ( 100 , 3 ) )
test_data = data.frame ( cbind ( id = 5:6 , dim = 3 , t ( x ), t ( y ) , t ( z ) ) )  

#Predict

predict.mclassiKernel ( object = object, newdata= test_data, predict.type = "response" )
}
